fixed migration

+++ database/migrations/2026_02_09_225053_add_can_edit_to_role_client_permissions_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -21,8 +20,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('role_client_permissions', function (Blueprint $table) {
-            $table->dropColumn('can_edit');
-        });
+        if (Schema::hasTable('role_client_permissions')) {
+            Schema::table('role_client_permissions', function (Blueprint $table) {
+                $table->dropColumn('can_edit');
+            });
+        }
     }
 };
